[MAC-MEMO]:폴더 삭제 기능구현   - 폴더 삭제 시 안의 메모까지 삭제기능 구현   - 폴더 삭제 시 폴더 리스트에 삭제한 폴더가 노출됨   - 불필요한 파일 삭제
