+++ mac-memo/src/db/data.json
@@ -19,14 +19,6 @@
     {
       "id": "9999",
       "name": "최근 삭제된 항목"
-    },
-    {
-      "id": "5",
-      "name": "안녕"
-    },
-    {
-      "id": "6",
-      "name": "새로운 폴더"
     }
   ],
   "memo": [
@@ -49,18 +41,6 @@
       "content": "안녕하세요11111234"
     },
     {
-      "id": "c73a",
-      "folder": "5",
-      "time": "2024-07-23T07:17:21.339Z",
-      "content": "11111123123"
-    },
-    {
-      "id": "559d",
-      "folder": "5",
-      "time": "2024-07-23T07:16:52.462Z",
-      "content": "안녕하세요1111"
-    },
-    {
       "id": "4bd8",
       "folder": "4",
       "time": "2024-07-23T08:30:58.153Z",
@@ -75,7 +55,7 @@
     {
       "id": "ea08",
       "folder": "3",
-      "time": "2024-07-23T08:49:41.598Z",
+      "time": "2024-07-24T01:49:08.296Z",
       "content": "asdasdasdasd123"
     },
     {
@@ -85,22 +65,10 @@
       "content": "1234"
     },
     {
-      "id": "653e",
-      "folder": "6",
-      "time": "2024-07-23T08:30:12.103Z",
-      "content": "안녕하세요"
-    },
-    {
       "id": "72d5",
       "folder": "3",
       "time": "2024-07-23T08:35:05.620Z",
       "content": "안녕하세요"
-    },
-    {
-      "id": "1de2",
-      "folder": "4",
-      "time": "2024-07-23T08:49:45.704Z",
-      "content": ""
     },
     {
       "id": "54f7",
@@ -109,16 +77,16 @@
       "content": "ㅇㅇㅇㅇㅇ"
     },
     {
-      "id": "e9bd",
-      "folder": "1",
-      "time": "2024-07-24T00:37:48.715Z",
-      "content": ""
+      "id": "87de",
+      "folder": "3",
+      "time": "2024-07-24T02:16:52.342Z",
+      "content": "123123"
     },
     {
-      "id": "62d9",
-      "folder": "1",
-      "time": "2024-07-24T00:38:03.778Z",
-      "content": ""
+      "id": "22b3",
+      "folder": "2",
+      "time": "2024-07-24T02:04:30.250Z",
+      "content": "123412"
     }
   ]
 }
